streaming store-gateway: more and rename loadingBatchSet
Renames `loadingBatchSet` to `loadingSeriesChunksSetIterator` and moves
it to `series_chunks.go`
