+++ pkg/storegateway/batch_series.go
@@ -91,77 +91,5 @@ func (s *BucketStore) batchSetsForBlocks(ctx context.Context, req *storepb.Serie
 }
 
 func newSeriesSetWithChunks(ctx context.Context, chunkReaders chunkReaders, chunksPool pool.Bytes, batches seriesChunkRefsSetIterator, stats *safeQueryStats) storepb.SeriesSet {
-	return newSeriesChunksSeriesSet(newPreloadingSeriesChunkSetIterator(ctx, 1, newLoadingBatchSet(chunkReaders, chunksPool, batches, stats)))
-}
-
-type loadingBatchSet struct {
-	chunkReaders chunkReaders
-	from         seriesChunkRefsSetIterator
-	chunksPool   pool.Bytes
-	stats        *safeQueryStats
-
-	current seriesChunksSet
-	err     error
-}
-
-func newLoadingBatchSet(chunkReaders chunkReaders, chunksPool pool.Bytes, from seriesChunkRefsSetIterator, stats *safeQueryStats) *loadingBatchSet {
-	return &loadingBatchSet{
-		chunkReaders: chunkReaders,
-		from:         from,
-		chunksPool:   chunksPool,
-		stats:        stats,
-	}
-}
-
-func (c *loadingBatchSet) Next() bool {
-	if c.err != nil {
-		return false
-	}
-
-	if !c.from.Next() {
-		c.err = c.from.Err()
-		return false
-	}
-
-	nextUnloaded := c.from.At()
-	entries := make([]seriesEntry, nextUnloaded.len())
-	c.chunkReaders.reset()
-	for i, s := range nextUnloaded.series {
-		entries[i].lset = s.lset
-		entries[i].chks = make([]storepb.AggrChunk, len(s.chunks))
-
-		for j, chunk := range s.chunks {
-			entries[i].chks[j].MinTime = chunk.minTime
-			entries[i].chks[j].MaxTime = chunk.maxTime
-
-			err := c.chunkReaders.addLoad(chunk.blockID, chunk.ref, i, j)
-			if err != nil {
-				c.err = errors.Wrap(err, "preloading chunks")
-				return false
-			}
-		}
-	}
-
-	// Create a batched memory pool that can be released all at once.
-	chunksPool := &pool.BatchBytes{Delegate: c.chunksPool}
-
-	err := c.chunkReaders.load(entries, chunksPool, c.stats)
-	if err != nil {
-		c.err = errors.Wrap(err, "loading chunks")
-		return false
-	}
-	nextLoaded := seriesChunksSet{
-		series:         entries,
-		chunksReleaser: chunksPool,
-	}
-	c.current = nextLoaded
-	return true
-}
-
-func (c *loadingBatchSet) At() seriesChunksSet {
-	return c.current
-}
-
-func (c *loadingBatchSet) Err() error {
-	return c.err
+	return newSeriesChunksSeriesSet(newPreloadingSeriesChunkSetIterator(ctx, 1, newLoadingSeriesChunksSetIterator(chunkReaders, chunksPool, batches, stats)))
 }
